// src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      ingresos: {
        Row: {
          id: number
          receipt_number: string
          dni: string
          full_name: string
          amount: number
          account: string
          date: string
          transaction_type: string
          created_at: string
          numeroOperacion: number | null
        }
        Insert: {
          id?: number
          receipt_number: string
          dni: string
          full_name: string
          amount: number
          account: string
          date: string
          transaction_type: string
          created_at?: string
          numeroOperacion?: number | null
        }
        Update: {
          id?: number
          receipt_number?: string
          dni?: string
          full_name?: string
          amount?: number
          account?: string
          date?: string
          transaction_type?: string
          created_at?: string
          numeroOperacion?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "ingresos_account_fkey"
            columns: ["account"]
            isOneToOne: false
            referencedRelation: "cuentas"
            referencedColumns: ["name"]
          },
          {
            foreignKeyName: "ingresos_dni_fkey"
            columns: ["dni"]
            isOneToOne: false
            referencedRelation: "socio_titulares"
            referencedColumns: ["dni"]
          }
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (Database['public']['Tables'] & Database['public']['Views'])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (Database['public']['Tables'] &
        Database['public']['Views'])
    ? (Database['public']['Tables'] &
        Database['public']['Views'])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never

export type TablesInsert<
  PublicTableNameOrOptions extends
    | keyof Database['public']['Tables']
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicTableNameOrOptions['schema']]['Tables']
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? Database[PublicTableNameOrOptions['schema']]['Tables'][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : PublicTableNameOrOptions extends keyof Database['public']['Tables']
    ? Database['public']['Tables'][PublicTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never